# Assignment IV/PizzeriaUnoServer/swagger.json
{
    "swagger":"2.0",
    "info":{
        "description":"The backend for the Pizza application",
        "version":"1.0.0",
        "title":"PizzeriaUno",
        "termsOfService":"http://swagger.io/terms/",
        "contact":{
            "email":"arnarl@ru.is"
        },
        "license":{
            "name":"Apache 2.0",
            "url":"http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host":"localhost:3503",
    "basePath":"/api",
    "schemes":[
        "http"
    ],
    "paths":{
        "/pizzas":{
            "get":{
                "tags":[
                    "Pizza"
                ],
                "summary":"Get all available pizzas",
                "description":"",
                "operationId":"getAvailablePizza",
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "responses":{
                    "200":{
                        "description":"A list of pizza objects",
                        "schema":{
                            "type":"array",
                            "properties":{
                                "id":{
                                    "type":"integer"
                                },
                                "name":{
                                    "type":"string"
                                },
                                "description":{
                                    "type":"string"
                                },
                                "price":{
                                    "type":"integer"
                                },
                                "image":{
                                    "type":"string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pizzas/{pizzaId}":{
            "get":{
                "tags":[
                    "Pizza"
                ],
                "summary":"Get pizza by id",
                "description":"",
                "operationId":"getPizzaById",
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters": [
                    { "name": "pizzaId", "in": "path", "description": "Id of the pizza to retrieve", "required": true, "type": "integer", "format": "int64" }
                ],
                "responses":{
                    "200":{
                        "description":"A single pizza object",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "id":{
                                    "type":"integer"
                                },
                                "name":{
                                    "type":"string"
                                },
                                "description":{
                                    "type":"string"
                                },
                                "price":{
                                    "type":"integer"
                                },
                                "image":{
                                    "type":"string"
                                }
                            }
                        }
                    },
                    "404": { "description": "Not found with provided id" }
                }
            }
        },
        "/offers":{
            "get":{
                "tags":[
                    "Offer"
                ],
                "summary":"Get all available offers",
                "description":"",
                "operationId":"getOffers",
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "responses":{
                    "200":{
                        "description":"A list of offer objects",
                        "schema":{
                            "type":"array",
                            "properties":{
                                "id":{
                                    "type":"integer"
                                },
                                "offer":{
                                    "type":"string"
                                },
                                "price":{
                                    "type":"integer"
                                },
                                "validFor":{
                                    "type":"string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/offers/{offerId}":{
            "get":{
                "tags":[
                    "Offer"
                ],
                "summary":"Get offer by id",
                "description":"",
                "operationId":"getOfferById",
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters": [
                    { "name": "offerId", "in": "path", "description": "Id of the offer to retrieve", "required": true, "type": "integer", "format": "int64" }
                ],
                "responses":{
                    "200":{
                        "description":"A single offer object",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "id":{
                                    "type":"integer"
                                },
                                "offer":{
                                    "type":"string"
                                },
                                "price":{
                                    "type":"integer"
                                },
                                "validFor":{
                                    "type":"string"
                                }
                            }
                        }
                    },
                    "404": { "description": "Not found with specified id" }
                }
            }
        },
        "/orders/{telephone}":{
            "get":{
                "tags":[
                    "Order"
                ],
                "summary":"Get order by telephone",
                "description":"",
                "operationId":"getOrderByTelephone",
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters": [
                    { "name": "telephone", "in": "path", "description": "Telephone of user associated with orders", "required": true, "type": "string" }
                ],
                "responses":{
                    "200":{
                        "description":"A single order object",
                        "schema":{
                            "type":"object"
                        }
                    }
                }
            },
            "post": {
                "tags": [ "Order" ],
                "summary": "Create a new order",
                "description": "",
                "consumes": [ "application/json" ],
                "produces": [ "application/json" ],
                "parameters": [
                    { "in": "body", "name": "body", "description": "The order object", "required": true }
                ],
                "response": {
                    "201": { "description": "An order was successfully created" },
                    "500": { "description": "A server-side error" }
                }
            }
        }
    }
}